optimize GlobalExceptionHandler and move CustomAuthenticationEntryPoint to identity module
