+++ app/src/main/kotlin/net/thechance/app/exception/GlobalExceptionHandler.kt
@@ -1,6 +1,5 @@
 package net.thechance.app.exception
 
-import net.thechance.app.exception.exceptions.*
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -11,16 +10,6 @@ import org.springframework.web.bind.annotation.RestControllerAdvice
 class GlobalExceptionHandler {
 
     private val logger = LoggerFactory.getLogger(GlobalExceptionHandler::class.java)
-
-    @ExceptionHandler(ResourceNotFoundException::class)
-    fun handleResourceNotFound(ex: ResourceNotFoundException): ResponseEntity<ApiErrorResponse> {
-        return createErrorResponse(ex, HttpStatus.NOT_FOUND.value(), "Resource Not Found")
-    }
-
-    @ExceptionHandler(ResourceConflictException::class)
-    fun handleResourceConflict(ex: ResourceConflictException): ResponseEntity<ApiErrorResponse> {
-        return createErrorResponse(ex, HttpStatus.CONFLICT.value(), "Resource Conflict")
-    }
 
     @ExceptionHandler(AccessForbiddenException::class)
     fun handleAccessForbidden(ex: AccessForbiddenException): ResponseEntity<ApiErrorResponse> {
@@ -38,12 +27,8 @@ class GlobalExceptionHandler {
     }
 
     @ExceptionHandler(Exception::class)
-    fun handleGeneral(): ResponseEntity<ApiErrorResponse> {
-        val apiError = ApiErrorResponse(
-            status = HttpStatus.INTERNAL_SERVER_ERROR.value(),
-            message = "An unexpected error occurred while processing your request"
-        )
-        return ResponseEntity(apiError, HttpStatus.INTERNAL_SERVER_ERROR)
+    fun handleGeneral(ex: Exception): ResponseEntity<ApiErrorResponse> {
+        return createErrorResponse(ex, HttpStatus.INTERNAL_SERVER_ERROR.value(), "Unexpected Error")
     }
 
     private fun createErrorResponse(
