+++ app/src/main/kotlin/net/thechance/app/exception/ValidationExceptions.kt
@@ -1,4 +1,8 @@
-package net.thechance.app.exception.exceptions
+package net.thechance.app.exception
+
+class AccessForbiddenException(
+    message: String = "You do not have permission to access this resource"
+) : RuntimeException(message)
 
 class RequestValidationException(
     message: String = "Request validation failed"
