+++ identity/src/main/kotlin/identity/security/handler/CustomAuthenticationEntryPoint.kt
@@ -1,8 +1,9 @@
-package net.thechance.app.exception
+package net.thechance.identity.security.handler
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
+import net.thechance.app.exception.ApiErrorResponse
 import org.springframework.security.core.AuthenticationException
 import org.springframework.security.web.AuthenticationEntryPoint
 import org.springframework.stereotype.Component
